Split CSS and update main and subpages
Split the CSS to be more manageable between a style for the main page and a style for the subpages. Updated the main page to include a foot with links to my GitHub and LinkedIn accounts. Updated the Athena subpage with some more details on the project and a link to it